Resolving Issue #2219
This new change will keep the host entry in the host table for a day
after the device has been removed.

+++ lib/api_device.php
@@ -45,7 +45,12 @@ function api_device_remove($device_id) {
 
 	api_plugin_hook_function('device_remove', array($device_id));
 
-	db_execute_prepared('DELETE FROM host             WHERE      id = ?', array($device_id));
+	if ($poller_id == 1) {
+		db_execute_prepared('DELETE FROM host         WHERE      id = ?', array($device_id));
+	} else {
+		db_execute_prepared('UPDATE host SET deleted = "on", poller_id = 1 WHERE id = ?', array($device_id));
+	}
+
 	db_execute_prepared('DELETE FROM host_graph       WHERE host_id = ?', array($device_id));
 	db_execute_prepared('DELETE FROM host_snmp_query  WHERE host_id = ?', array($device_id));
 	db_execute_prepared('DELETE FROM host_snmp_cache  WHERE host_id = ?', array($device_id));
@@ -91,6 +96,19 @@ function api_device_purge_from_remote($device_id, $poller_id = 0) {
 			db_execute_prepared('DELETE FROM poller_command   WHERE command LIKE ?', array($device_id . ':%'), true, $rcnn_id);
 			db_execute_prepared('DELETE FROM data_local       WHERE host_id = ?', array($device_id), true, $rcnn_id);
 			db_execute_prepared('DELETE FROM graph_local      WHERE host_id = ?', array($device_id), true, $rcnn_id);
+		}
+	}
+}
+
+function api_device_purge_deleted_devices() {
+	$devices = db_fetch_assoc_prepared('SELECT id
+		FROM host
+		WHERE deleted = "on"
+		AND UNIX_TIMESTAMP(last_updated) < UNIX_TIMESTAMP()-86400');
+
+	if (cacti_sizeof($devices)) {
+		foreach($devices as $d) {
+			api_device_remove($d['id']);
 		}
 	}
 }
@@ -144,7 +162,10 @@ function api_device_remove_multi($device_ids, $delete_type = 2) {
 
 		$poller_ids = get_remote_poller_ids_from_devices($devices_to_delete);
 
-		db_execute("DELETE FROM host             WHERE id IN ($devices_to_delete)");
+		// handle removal or mark for removal as required
+		db_execute("DELETE FROM host WHERE id IN ($devices_to_delete) AND poller_id = 1");
+		db_execute("UPDATE host SET deleted = 'on', poller_id = 1 WHERE id IN ($devices_to_delete) AND poller_id != 1");
+
 		db_execute("DELETE FROM host_graph       WHERE host_id IN ($devices_to_delete)");
 		db_execute("DELETE FROM host_snmp_query  WHERE host_id IN ($devices_to_delete)");
 		db_execute("DELETE FROM host_snmp_cache  WHERE host_id IN ($devices_to_delete)");
@@ -156,8 +177,8 @@ function api_device_remove_multi($device_ids, $delete_type = 2) {
 		} else {
 			api_data_source_disable_multi($data_sources);
 
-			db_execute("UPDATE graph_local SET host_id=0 WHERE host_id IN($devices_to_delete)");
-			db_execute("UPDATE data_local SET host_id=0 WHERE host_id IN($devices_to_delete)");
+			db_execute("UPDATE graph_local SET host_id = 0 WHERE host_id IN($devices_to_delete)");
+			db_execute("UPDATE data_local  SET host_id = 0 WHERE host_id IN($devices_to_delete)");
 		}
 
 		if (cacti_sizeof($poller_ids)) {
